`model_dtype` initial

+++ nequip/scripts/train.py
@@ -34,8 +34,9 @@ default_config = dict(
         "RescaleEnergyEtc",
     ],
     dataset_statistics_stride=1,
-    default_dtype="float32",
-    allow_tf32=False,  # TODO: until we understand equivar issues
+    default_dtype="float64",
+    model_dtype="float32",
+    allow_tf32=False,
     verbose="INFO",
     model_debug_mode=False,
     equivariance_test=False,
